+++ src/shared/components/display/TextWithIconBlock.tsx
